+++ spec/assessments/outboundLinksSpec.js
@@ -30,7 +30,7 @@ describe( "An assessor running the linkStatistics", function() {
 		var assessment = linkStatisticAssessment.getResult( mockPaper, factory.buildMockResearcher( mockResult ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 8 );
-		expect( assessment.getText() ).toEqual( "This page has 0 nofollowed outbound link(s) and 1 normal outbound link(s)." );
+		expect( assessment.getText() ).toEqual( "This page has 0 nofollowed <a href='https://yoa.st/2pl' target='_blank'>outbound link(s)</a> and 1 normal outbound link(s)." );
 
 		mockPaper = new Paper( "a test with a <a href='http://yoast.com' alt='' rel='nofollow'> link </a>", attributes );
 
@@ -50,7 +50,7 @@ describe( "An assessor running the linkStatistics", function() {
 		assessment = linkStatisticAssessment.getResult( mockPaper, factory.buildMockResearcher( mockResult ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 7 );
-		expect( assessment.getText() ).toEqual( "This page has 1 outbound link(s), all nofollowed." );
+		expect( assessment.getText() ).toEqual( "This page has 1 <a href='https://yoa.st/2pl' target='_blank'>outbound link(s)</a>, all nofollowed." );
 	} );
 
 	it( "Accepts a paper and i18nobject  ", function() {
@@ -58,6 +58,6 @@ describe( "An assessor running the linkStatistics", function() {
 		var assessment = linkStatisticAssessment.getResult( mockPaper, factory.buildMockResearcher( { externalTotal: 0 } ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 6 );
-		expect( assessment.getText() ).toEqual( "No outbound links appear in this page, consider adding some as appropriate." );
+		expect( assessment.getText() ).toEqual( "No <a href='https://yoa.st/2pl' target='_blank'>outbound links</a> appear in this page, consider adding some as appropriate." );
 	} );
 } );