+++ src/assessments/readability/fleschReadingEaseAssessment.js
@@ -4,7 +4,7 @@ const inRange = require( "lodash/inRange" );
 
 const getLanguageAvailability = require( "../../helpers/getLanguageAvailability.js" );
 
-const availableLanguages = [ "en", "nl", "de", "it", "ru" ];
+const availableLanguages = [ "en", "nl", "de", "it", "ru", "fr", "es" ];
 
 class FleschReadingEaseAssessment extends Assessment {
 	/**