+++ src/researches/findKeywordInFirstParagraph.js
@@ -1,11 +1,11 @@
 /** @module analyses/findKeywordInFirstParagraph */
 
-var matchParagraphs = require( "../stringProcessing/matchParagraphs.js" );
-var wordMatch = require( "../stringProcessing/matchTextWithWord.js" );
+const matchParagraphs = require( "../stringProcessing/matchParagraphs.js" );
+const wordMatch = require( "../stringProcessing/matchTextWithWord.js" );
 
-var escapeRegExp = require( "lodash/escapeRegExp" );
-var reject = require( "lodash/reject" );
-var isEmpty = require( "lodash/isEmpty" );
+const escapeRegExp = require( "lodash/escapeRegExp" );
+const reject = require( "lodash/reject" );
+const isEmpty = require( "lodash/isEmpty" );
 
 /**
  * Counts the occurrences of the keyword in the first paragraph, returns 0 if it is not found,
@@ -16,8 +16,8 @@ var isEmpty = require( "lodash/isEmpty" );
  * @returns {number} The number of occurrences of the keyword in the first paragraph.
  */
 module.exports = function( paper ) {
-	var paragraphs = matchParagraphs( paper.getText() );
-	var keyword = escapeRegExp( paper.getKeyword().toLocaleLowerCase() );
-	var paragraph = reject( paragraphs, isEmpty )[ 0 ] || "";
-	return wordMatch( paragraph, keyword, paper.getLocale() );
+	const paragraphs = matchParagraphs( paper.getText() );
+	const keyword = escapeRegExp( paper.getKeyword().toLocaleLowerCase() );
+	const paragraph = reject( paragraphs, isEmpty )[ 0 ] || "";
+	return wordMatch( paragraph, keyword, paper.getLocale() ).count;
 };