+++ src/stringProcessing/matchTextWithWord.js
@@ -1,8 +1,12 @@
 /** @module stringProcessing/matchTextWithWord */
 
-var stripSomeTags = require( "../stringProcessing/stripNonTextTags.js" );
-var unifyWhitespace = require( "../stringProcessing/unifyWhitespace.js" ).unifyAllSpaces;
-var matchStringWithTransliteration = require( "../stringProcessing/matchTextWithTransliteration.js" );
+const stripSomeTags = require( "../stringProcessing/stripNonTextTags.js" );
+const stripSpaces = require( "../stringProcessing/stripSpaces.js" );
+const removePunctuation = require( "../stringProcessing/removePunctuation.js" );
+const unifyWhitespace = require( "../stringProcessing/unifyWhitespace.js" ).unifyAllSpaces;
+const matchStringWithTransliteration = require( "../stringProcessing/matchTextWithTransliteration.js" );
+const normalizeQuotes = require( "../stringProcessing/quotes.js" ).normalize;
+const map = require( "lodash/map" );
 
 /**
  * Returns the number of matches in a given string
@@ -11,11 +15,20 @@ var matchStringWithTransliteration = require( "../stringProcessing/matchTextWith
  * @param {string} wordToMatch The word to match in the text
  * @param {string} locale The locale used for transliteration.
  * @param {string} [extraBoundary] An extra string that can be added to the wordboundary regex
- * @returns {number} The amount of matches found.
+ * @returns {Object} The matches found and the number of matches.
  */
 module.exports = function( text, wordToMatch, locale, extraBoundary ) {
 	text = stripSomeTags( text );
 	text = unifyWhitespace( text );
-	var matches = matchStringWithTransliteration( text, wordToMatch, locale, extraBoundary );
-	return matches.length;
+	text = normalizeQuotes( text );
+	wordToMatch = normalizeQuotes( wordToMatch );
+	let matches = matchStringWithTransliteration( text, wordToMatch, locale, extraBoundary );
+	matches = map( matches, function( keyword ) {
+		return stripSpaces( removePunctuation( keyword ) );
+	} );
+
+	return {
+		count: matches.length,
+		matches: matches,
+	};
 };