+++ CHANGELOG.md
@@ -5,6 +5,26 @@ This changelog is according to [Keep a Changelog](http://keepachangelog.com).
 All notable changes to this project will be documented in this file.
 We will follow [Semantic Versioning](http://semver.org/) from version 2 and onwards.
 
+## 1.35.0 July 3rd, 2018
+### Added
+* Adds an assessment the checks the distribution of the focus keyword in the text.
+* Adds a topic distribution and topic density assessment to use with synonyms.
+* Adds links to relevant articles to all SEO assessments.
+* Adds Flesch Reading Ease assessment for French.
+* Adds Flesch Reading Ease assessment for Spanish.
+* Add passive voice assessment for Italian.
+
+### Fixed
+* Instances of the same keyword with different kinds of apostrophes (e.g., brain’s and brain's) are now recognized as the same in the following assessments: keyword in meta description, keyword in subheading, keyword in first paragraph, keyword in title, keyword in URL.
+* Filters out prominent word combinations ending in 's in English. Props to [swekkiekekkie](https://github.com/swekkiekekkie).
+
+## 1.34.0 June 12th, 2018
+### Fixed
+* Fixes title width measurements by adding font styles to the hidden input field.
+
+### Changed
+* Improves the lists of French transition words, stopwords, and function words, props [Laurent-1971](https://github.com/Laurent-1971).
+
 ## 1.33.1 May 25th, 2018
 ### Fixed
 * Fixes a bug where Flesch Reading Ease translation strings were not fully translated.